Refactor multiple classes Ensure consistent variable naming Ensure appropriate imports
