+++ src/main/java/fittrack/trainingsession/TrainingSession.java
@@ -1,4 +1,6 @@
-package seedu.duke;
+package fittrack.trainingsession;
+
+import fittrack.exercisestation.*;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -25,14 +27,14 @@ public class TrainingSession{
     private ExerciseStation[] exerciseData = {new PullUpStation(), new ShuttleRunStation(), new SitAndReachStation(),
             new SitUpStation(), new StandingBroadJumpStation(), new WalkAndRunStation()};
 
-    TrainingSession(String datetime, String sessionDescription){
+    public TrainingSession(String datetime, String sessionDescription){
         this.sessionDatetime = LocalDateTime.parse(datetime + "-ss-ns");
         this.sessionDescription = sessionDescription;
     }
 
     //Edits session data
     public void editExercise(int exerciseNum, int reps){
-        exerciseData[exerciseNum].setReps(reps);
+        exerciseData[exerciseNum].setPerformance(reps);
         System.out.print("Exercise edited! Here's your new input: " + System.lineSeparator() +
                 exerciseData[exerciseNum]);
     }
